fix(file-manager): open supported files with the expected preview
