+++ packages/workspace/file-manager/src/i18n/workspaceFileManagerI18n.ts
@@ -47,6 +47,8 @@ const workspaceFileManagerEn = {
   downloadLabel: "Download",
   dropToImportLabel: "Drop files here to import to the current folder",
   emptyDirectory: "No visible files in this folder",
+  collapseFolderLabel: "Collapse folder",
+  expandFolderLabel: "Expand folder",
   forwardLabel: "Forward",
   loading: "Loading",
   modifiedLabel: "Modified",
@@ -149,6 +151,8 @@ const workspaceFileManagerZhCN = {
   downloadLabel: "下载",
   dropToImportLabel: "将文件拖到这里以导入到当前文件夹",
   emptyDirectory: "这个文件夹里没有可见文件",
+  collapseFolderLabel: "收起文件夹",
+  expandFolderLabel: "展开文件夹",
   forwardLabel: "前进",
   loading: "加载中",
   modifiedLabel: "修改时间",
