+++ packages/workspace/file-manager/src/services/internal/model/entryLookup.ts
@@ -1,0 +1,29 @@
+import type {
+  WorkspaceFileEntry,
+  WorkspaceFileManagerState
+} from "../../workspaceFileManagerTypes.ts";
+
+export function findWorkspaceFileEntry(
+  state: WorkspaceFileManagerState,
+  entryPath: string | null | undefined
+): WorkspaceFileEntry | null {
+  if (!entryPath) {
+    return null;
+  }
+
+  const directEntry = state.entries.find((entry) => entry.path === entryPath);
+  if (directEntry) {
+    return directEntry;
+  }
+
+  for (const directoryState of Object.values(state.directoryExpansionByPath)) {
+    const nestedEntry = directoryState.entries.find(
+      (entry) => entry.path === entryPath
+    );
+    if (nestedEntry) {
+      return nestedEntry;
+    }
+  }
+
+  return null;
+}
